+++ packages/frontend-2/lib/common/composables/async.ts
@@ -1,25 +1,10 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { MaybeAsync } from '@speckle/shared'
-import { AsyncComputedOptions, computedAsync } from '@vueuse/core'
+import {
+  AsyncWritableComputedOptions,
+  AsyncWritableComputedRef,
+  writableAsyncComputed as originalWritableAsyncComputed
+} from '@speckle/ui-components'
 
-export interface AsyncWritableComputedOptions<T> {
-  get: (...args: any[]) => MaybeAsync<T>
-  set: (value: T) => MaybeAsync<void>
-  initialState: T
-  readOptions?: AsyncComputedOptions
-  asyncRead?: boolean
-  debugging?: Partial<{
-    log: {
-      name: string
-      writesOnly?: boolean
-      readsOnly?: boolean
-    }
-  }>
-}
-
-export type AsyncWritableComputedRef<T> = ComputedRef<T> & {
-  update: AsyncWritableComputedOptions<T>['set']
-}
+export type { AsyncWritableComputedOptions, AsyncWritableComputedRef }
 
 /**
  * Allows async read/write from/to computed. Use `res.value` to read and `res.update` to write. If you only need
@@ -27,41 +12,18 @@ export type AsyncWritableComputedRef<T> = ComputedRef<T> & {
  * disable async reads through the `asyncRead` param.
  * @param params
  */
-export function writableAsyncComputed<T>(
-  params: AsyncWritableComputedOptions<T>
-): AsyncWritableComputedRef<T> {
-  const { get, initialState, readOptions, set, asyncRead = true, debugging } = params
-  const logSettings = debugging?.log
-  const getTrace = () => (new Error('Trace:').stack || '').substring(7)
+export const writableAsyncComputed: typeof originalWritableAsyncComputed = (params) => {
   const logger = useLogger()
-
-  const finalGet: typeof get =
-    logSettings && !logSettings.writesOnly
-      ? () => {
-          const res = get()
-          logger.debug(`debugging: '${logSettings.name}' read`, res, getTrace())
-          return res
+  return originalWritableAsyncComputed({
+    ...params,
+    debugging: params.debugging?.log
+      ? {
+          ...params.debugging,
+          log: {
+            ...params.debugging.log,
+            logger: logger.debug
+          }
         }
-      : get
-
-  const finalSet: typeof set =
-    logSettings && !logSettings.readsOnly
-      ? (newVal) => {
-          logger.debug(
-            `debugging: '${logSettings.name}' written to`,
-            newVal,
-            getTrace()
-          )
-          return set(newVal)
-        }
-      : set
-
-  const readValue = asyncRead
-    ? computedAsync(finalGet, initialState, readOptions)
-    : computed(finalGet)
-
-  const getter = computed(() => readValue.value) as AsyncWritableComputedRef<T>
-  getter.update = finalSet
-
-  return getter
+      : undefined
+  })
 }
